+++ util/pkg/vfs/acls/gce/storage.go
@@ -22,11 +22,11 @@ import (
 
 	storage "google.golang.org/api/storage/v1"
 	"k8s.io/klog/v2"
-	"k8s.io/kops/pkg/acls"
 	"k8s.io/kops/pkg/apis/kops"
 	"k8s.io/kops/upup/pkg/fi/cloudup"
 	"k8s.io/kops/upup/pkg/fi/cloudup/gce"
 	"k8s.io/kops/util/pkg/vfs"
+	"k8s.io/kops/util/pkg/vfs/acls"
 )
 
 // gcsAclStrategy is the AclStrategy for objects written to google cloud storage
